+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 # Changes
 - Fix some bugs in buttsbot
+- Add command line option to save file with a command list